Spec.DynamicLogic.Counters

Description

Define several variant models of counters which are useful to test or use examples for various behaviours of the runtime.

Documentation

newtype SimpleCounter

Constructors

SimpleCounter

Fields

count :: Int

newtype FailingCounter

Constructors

FailingCounter

Fields

failingCount :: Int

data Counter

Constructors

Counter Int