Test.QuickCheck.DynamicLogic.CanGenerate

Documentation

canGenerate :: Double -> Gen a -> (a -> Bool) -> Bool

canGenerate prob g p returns False if we are sure Prob(g generates x satisfying p) >= prob otherwise True (and we know such an x can be generated).