Test.QuickCheck.DynamicLogic.Monad

Documentation

data DL s a

The DL monad provides a nicer interface to dynamic logic formulae than the plain API. It's a continuation monad producing a DynFormula formula, with a state component threaded through.

action :: (Show a, Typeable a, Eq (Action s a)) => Action s a -> DL s ()

anyAction :: DL s ()

anyActions :: Int -> DL s ()

anyActions_ :: DL s ()

stopping :: DL s ()

weight :: Double -> DL s ()

getSize :: DL s Int

getModelStateDL :: DL s s

assert :: String -> Bool -> DL s ()

Fail if the boolean is False.

Equivalent to

assert msg b = unless b (fail msg)

assertModel :: String -> (s -> Bool) -> DL s ()

monitorDL :: (Property -> Property) -> DL s ()

forAllQ :: Quantifiable q => q -> DL s (Quantifies q)

Generate a random value using the given Quantification (or list/tuple of quantifications). Generated values will only shrink to smaller values that could also have been generated.

forAllDL :: (DynLogicModel s, Testable a) => DL s () -> (Actions s -> a) -> Property

forAllDL_ :: (DynLogicModel s, Testable a) => DL s () -> (Actions s -> a) -> Property

forAllMappedDL :: (DynLogicModel s, Testable a, Show rep) => (rep -> DynLogicTest s) -> (DynLogicTest s -> rep) -> (Actions s -> srep) -> DL s () -> (srep -> a) -> Property

forAllMappedDL_ :: (DynLogicModel s, Testable a, Show rep) => (rep -> DynLogicTest s) -> (DynLogicTest s -> rep) -> (Actions s -> srep) -> DL s () -> (srep -> a) -> Property

forAllUniqueDL :: (DynLogicModel s, Testable a) => Int -> s -> DL s () -> (Actions s -> a) -> Property

withDLTest :: (DynLogicModel s, Testable a) => DL s () -> (Actions s -> a) -> DynLogicTest s -> Property

data DynLogic s

Base Dynamic logic formulae language. Formulae are parameterised over the type of state s to which they apply. A DynLogic value cannot be constructed directly, one has to use the various "smart constructors" provided, see the Building formulae section.

class StateModel s => DynLogicModel s where

Restricted calls are not generated by AfterAny; they are included in tests explicitly using After in order to check specific properties at controlled times, so they are likely to fail if invoked at other times.

Minimal complete definition

Nothing

Methods

restricted :: Action s a -> Bool

data DynLogicTest s

Constructors

BadPrecondition [TestStep s] [Any (Action s)] s

Looping [TestStep s]

Stuck [TestStep s] s

DLScript [TestStep s]

data TestStep s

Constructors

Do (Step s)

forall a.(Eq a, Show a, Typeable a) => Witness a

module Test.QuickCheck.DynamicLogic.Quantify