Test.QuickCheck.StateModel.Variables

Documentation

data Var a

A symbolic variable for a value of type a

data Any f where

Constructors

Some :: (Typeable a, Eq (f a)) => f a -> Any f

class HasVariables a where

This type class gives you a way to get all the symbolic variables that appear in a value.

Methods

getAllVariables :: a -> Set (Any Var)

newtype HasNoVariables a

Constructors

HasNoVariables a

data VarContext

mkVar :: Int -> Var a

Create a fresh symbolic variable with given identifier. While Vars are usually created by action generators, this function can be used for example to create a Var in the initialState of a StateModel. A good default value for the identifier is '-1' as this will not be generated otherwise.

ctxAtType :: Typeable a => VarContext -> [Var a]

arbitraryVar :: Typeable a => VarContext -> Gen (Var a)

shrinkVar :: Typeable a => VarContext -> Var a -> [Var a]

extendContext :: Typeable a => VarContext -> Var a -> VarContext

isWellTyped :: Typeable a => Var a -> VarContext -> Bool

allVariables :: HasVariables a => a -> VarContext

isEmptyCtx :: VarContext -> Bool

unsafeCoerceVar :: Var a -> Var b

unsafeNextVarIndex :: VarContext -> Int