Test.QuickCheck.DynamicLogic.Quantify

Description

This module defines Quantifications, which are used together with forAllQ in DynamicLogic. A `Quantification t` can be used to generate a t, shrink a t, and recognise a generated t.

Documentation

data Quantification a

A Quantification over a type a is a generator that can be used to generate random values in DL scenarios.

A Quantification is similar to a Arbitrary, it groups together:

A standard QuickCheck _generator_ in the Gen monad, which can be "empty",

A _shrinking_ strategy for generated values in the case of a failures ensuring they stay within the domain,

A _predicate_ allowing finer grained control on generation and shrinking process, e.g in the case the range of the generator depends on trace context.

NOTE: Leaving the possibility of generating Nothing is useful to simplify the generation process for elements or frequency which may normally crash when the list to select elements from is empty. This makes writing DL formulas cleaner, removing the need to handle non-existence cases explicitly.

type QuantifyConstraints a = (Eq a, Show a, Typeable a, HasVariables a)

isEmptyQ :: Quantification a -> Bool

generateQ :: Quantification a -> Gen a

shrinkQ :: Quantification a -> a -> [a]

arbitraryQ :: Arbitrary a => Quantification a

Pack up an Arbitrary instance as a Quantification. Treats all values as being in range.

exactlyQ :: Eq a => a -> Quantification a

Generates exactly the given value. Does not shrink.

elementsQ :: Eq a => [a] -> Quantification a

Pick a random value from a list. Treated as an empty choice if the list is empty:

forAllQ (elementsQ []) == empty

oneofQ :: [Quantification a] -> Quantification a

Choose from a list of quantifications. Same as frequencyQ with all weights the same (and > 0).

frequencyQ :: [(Int, Quantification a)] -> Quantification a

Choose from a weighted list of quantifications. Treated as an empty choice if no quantification has weight > 0.

mapQ :: (a -> b, b -> a) -> Quantification a -> Quantification b

Quantification is not a Functor, since it also keeps track of the range of the generators. However, if you have two functions to :: a -> b from :: b -> a satisfying from . to = id you can go from a quantification over a to one over b. Note that the from function need only be defined on the image of to.

whereQ :: Quantification a -> (a -> Bool) -> Quantification a

Restrict the range of a quantification.

chooseQ :: (Arbitrary a, Random a, Ord a) => (a, a) -> Quantification a

Generate a random value in a given range (inclusive).

withGenQ :: Gen a -> (a -> Bool) -> (a -> [a]) -> Quantification a

Construct a `Quantification a` from its constituents. Note the predicate provided is used to restrict both the range of values generated and the list of possible shrinked values.

hasNoVariablesQ :: Quantification a -> Quantification (HasNoVariables a)

Wrap a Quantification in HasNoVariables to indicate that you know what you're doing and there are no symbolic variables in the thing you are quantifying over. WARNING: use this function carefully as there is no guarantee that you won't get bitten by very strange failures if you were in fact not honest about the lack of variables.

validQuantification :: Show a => Quantification a -> Property

Turns a Quantification into a Property to enable QuickChecking its validity.

class QuantifyConstraints (Quantifies q) => Quantifiable q where

Generalization of Quantifications, which lets you treat lists and tuples of quantifications as quantifications. For instance,

  ...
  (die1, die2) <- forAllQ (chooseQ (1, 6), chooseQ (1, 6))
  ...

Associated Types

type Quantifies q

The type of values quantified over.

Quantifies (Quantification a) = a

Methods

quantify :: q -> Quantification (Quantifies q)

Computing the actual Quantification.