Test.QuickCheck.StateModel

Description

Model-Based Testing library for use with Haskell QuickCheck.

This module provides the basic machinery to define a StateModel from which traces can be generated and executed against some actual implementation code to define monadic Property to be asserted by QuickCheck.

Documentation

module Test.QuickCheck.StateModel.Variables

class (forall a. Show (Action state a), forall a. HasVariables (Action state a), Show state, HasVariables state) => StateModel state where

The typeclass users implement to define a model against which to validate some implementation.

To implement a StateModel, user needs to provide at least the following:

A datatype for Actions: Each test case is a sequence of Actions that's supposed to lead from some initialState to some end state,

A generator for traces of Actions, the arbitraryAction function,

An initialState,

A transition function, nextState, that "interprets" each Action and producing some new state.

For finer grained control over the testing process, one can also define:

shrinkAction: Shrinking is an important part of MBT as it allows QuickCheck engine to look for simpler test cases when something goes wrong which makes troubleshooting easier,

precondition: Filters generated Action depending on the state. When precondition is False then the action is rejected and a new one is tried. This is also useful when shrinking a trace in order to ensure that removing some Action still produces a valid trace. The precondition can be somewhat redundant with the generator's conditions,

validFailingAction: Specifies when an action that fails it's precondition can still run as what is called a _negative_ action. This means that the action is (1) expected to fail and (2) not expected to change the model state. This is very useful for testing the checks and failure conditions in the SUT are implemented correctly. Should it be necessary to update the model state with e.g. book-keeping for a negative action one can define failureNextState - but it is generally recommended to let this be as simple an action as possible.

Minimal complete definition

arbitraryAction, initialState

Associated Types

data Action state a

The type of Action relevant for this state.

This is expected to be defined as a GADT where the a parameter is instantiated to some observable output from the SUT a given action is expected to produce. For example, here is a fragment of the `Action RegState` (taken from the RegistryModel module) :

  data Action RegState a where
    Spawn      ::                           Action RegState ThreadId
    Register   :: String -> Var ThreadId -> Action RegState ()
    KillThread :: Var ThreadId           -> Action RegState ()

The Spawn action should produce a ThreadId, whereas the KillThread action does not return anything.

Methods

actionName :: Action state a -> String

Display name for Action. This is useful to provide sensible statistics about the distribution of Actions run when checking a property.

Default implementation uses a poor-man's string manipulation method to extract the constructor name from the value.

arbitraryAction :: VarContext -> state -> Gen (Any (Action state))

Generator for Action depending on state.

shrinkAction :: Typeable a => VarContext -> state -> Action state a -> [Any (Action state)]

Shrinker for Action. Defaults to no-op but as usual, defining a good shrinker greatly enhances the usefulness of property-based testing.

initialState :: state

Initial state of generated traces.

nextState :: Typeable a => state -> Action state a -> Var a -> state

Transition function for the model. The `Var a` parameter is useful to keep reference to actual value of type a produced by performing the Action inside the state so that further actions can use Lookup to retrieve that data. This allows the model to be ignorant of those values yet maintain some references that can be compared and looked for.

failureNextState :: Typeable a => state -> Action state a -> state

Transition function for negative actions. Note that most negative testing applications should not require an implementation of this function!

precondition :: state -> Action state a -> Bool

Precondition for filtering generated Action. This function is applied before the action is performed, it is useful to refine generators that can produce more values than are useful.

validFailingAction :: state -> Action state a -> Bool

Precondition for filtering an Action that can meaningfully run but is supposed to fail. An action will run as a _negative_ action if the precondition fails and validFailingAction succeeds. A negative action should have _no effect_ on the model state. This may not be desierable in all situations - in which case one can override this semantics for book-keeping in failureNextState.

class (forall a. Show (Action state a), Monad m) => RunModel state m where

Minimal complete definition

perform

Associated Types

type Error state m

The type of errors that actions can throw. If this is defined as anything other than Void perform is required to return `Either (Error state) a` instead of a.

type Error state m = Void

Methods

perform :: Typeable a => state -> Action state a -> LookUp -> m (PerformResult state m a)

Perform an Action in some state in the Monad m. This is the function that's used to exercise the actual stateful implementation, usually through various side-effects as permitted by m. It produces a value of type a, eg. some observable output from the Action that should later be kept in the environment through a `Var a` also passed to the nextState function.

The Lookup parameter provides an environment to lookup `Var a` instances from previous steps.

postcondition :: (state, state) -> Action state a -> LookUp -> a -> PostconditionM m Bool

Postcondition on the a value produced at some step. The result is asserted and will make the property fail should it be False. This is useful to check the implementation produces expected values.

postconditionOnFailure :: (state, state) -> Action state a -> LookUp -> Either (Error state m) a -> PostconditionM m Bool

Postcondition on the result of running a _negative_ Action. The result is asserted and will make the property fail should it be False. This is useful to check the implementation produces e.g. the expected errors or to check that the SUT hasn't been updated during the execution of the negative action.

monitoring :: (state, state) -> Action state a -> LookUp -> Either (Error state m) a -> Property -> Property

Allows the user to attach additional information to the Property at each step of the process. This function is given the full transition that's been executed, including the start and ending state, the Action, the current environment to Lookup and the value produced by perform while executing this step.

monitoringFailure :: state -> Action state a -> LookUp -> Error state m -> Property -> Property

Allows the user to attach additional information to the Property if a positive action fails.

newtype PostconditionM m a

Constructors

PostconditionM

Fields

runPost :: WriterT (Endo Property, Endo Property) m a

data WithUsedVars a

Constructors

WithUsedVars VarContext a

data Annotated state

Constructors

Metadata

Fields

vars :: VarContext

underlyingState :: state

data Step state where

Constructors

(:=) :: (Typeable a, Eq (Action state a), Show (Action state a)) => Var a -> ActionWithPolarity state a -> Step state infix 5

data Polarity

Constructors

PosPolarity

NegPolarity

data ActionWithPolarity state a

Constructors

Eq (Action state a) => ActionWithPolarity

Fields

polarAction :: Action state a

polarity :: Polarity

type LookUp = forall a. Typeable a => Var a -> a

data Actions state

Constructors

Actions_ [String] (Smart [Step state])

pattern Actions :: [Step state] -> Actions state

data EnvEntry where

Constructors

(:==) :: Typeable a => Var a -> a -> EnvEntry infix 5

pattern (:=?) :: forall a. Typeable a => Var a -> a -> EnvEntry

type Env = [EnvEntry]

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to

class IsPerformResult e a

Minimal complete definition

performResultToEither

data Options state

Constructors

Options

Fields

actionLengthMultiplier :: Rational

monitorPost :: Monad m => (Property -> Property) -> PostconditionM m ()

Apply the property transformation to the property after evaluating the postcondition. Useful for collecting statistics while avoiding duplication between monitoring and postcondition.

counterexamplePost :: Monad m => String -> PostconditionM m ()

Acts as counterexample if the postcondition fails.

stateAfter :: forall state. StateModel state => Actions state -> Annotated state

runActions :: forall state m e. (StateModel state, RunModel state m, e ~ Error state m, forall a. IsPerformResult e a) => Actions state -> PropertyM m (Annotated state, Env)

lookUpVar :: Typeable a => Env -> Var a -> a

lookUpVarMaybe :: forall a. Typeable a => Env -> Var a -> Maybe a

viewAtType :: forall a. Typeable a => EnvEntry -> Maybe (Var a, a)

initialAnnotatedState :: StateModel state => Annotated state

computeNextState :: (StateModel state, Typeable a) => Annotated state -> ActionWithPolarity state a -> Var a -> Annotated state

computePrecondition :: StateModel state => Annotated state -> ActionWithPolarity state a -> Bool

computeArbitraryAction :: StateModel state => Annotated state -> Gen (Any (ActionWithPolarity state))

computeShrinkAction :: forall state a. (Typeable a, StateModel state) => Annotated state -> ActionWithPolarity state a -> [Any (ActionWithPolarity state)]

generateActionsWithOptions :: forall state. StateModel state => Options state -> Gen (Actions state)

Generate arbitrary actions with the GenActionsOptions. More flexible than using the type-based modifiers.

shrinkActionsWithOptions :: forall state. StateModel state => Options state -> Actions state -> [Actions state]

defaultOptions :: Options state

moreActions :: QCDProp state p => Rational -> p -> QCDProperty state