Test.QuickCheck.StateModel.Variables

Documentation

data Var a

A symbolic variable for a value of type a

data Any f where

Constructors

Some :: (Typeable a, Eq (f a)) => f a -> Any f

class HasVariables a where

This type class gives you a way to get all the symbolic variables that appear in a value.

Minimal complete definition

Nothing

Methods

getAllVariables :: a -> Set (Any Var)

default getAllVariables :: (Generic a, GenericHasVariables (Rep a)) => a -> Set (Any Var)

newtype HasNoVariables a

Constructors

HasNoVariables a

data VarContext

mkVar :: Int -> Var a

ctxAtType :: Typeable a => VarContext -> [Var a]

arbitraryVar :: Typeable a => VarContext -> Gen (Var a)

shrinkVar :: Typeable a => VarContext -> Var a -> [Var a]

extendContext :: Typeable a => VarContext -> Var a -> VarContext

isWellTyped :: Typeable a => Var a -> VarContext -> Bool

allVariables :: HasVariables a => a -> VarContext

unsafeCoerceVar :: Var a -> Var b

unsafeNextVarIndex :: VarContext -> Int